// frontend2/src/utils/types/models/TournamentSubmission.ts
export interface TournamentSubmission {
    /**
     * 
     * @type {number}
     * @memberof TournamentSubmission
     */
    readonly id: number;
    /**
     * 
     * @type {StatusBccEnum}
     * @memberof TournamentSubmission
     */
    readonly status: StatusBccEnum;
    /**
     * 
     * @type {string}
     * @memberof TournamentSubmission
     */
    readonly logs: string;
    /**
     * 
     * @type {string}
     * @memberof TournamentSubmission
     */
    readonly episode: string;
    /**
     * 
     * @type {number}
     * @memberof TournamentSubmission
     */
    readonly team: number;
    /**
     * 
     * @type {string}
     * @memberof TournamentSubmission
     */
    readonly teamname: string;
    /**
     * 
     * @type {number}
     * @memberof TournamentSubmission
     */
    readonly user: number;
    /**
     * 
     * @type {string}
     * @memberof TournamentSubmission
     */
    readonly username: string;
    /**
     * 
     * @type {Date}
     * @memberof TournamentSubmission
     */
    readonly created: Date;
    /**
     * 
     * @type {boolean}
     * @memberof TournamentSubmission
     */
    readonly accepted: boolean;
    /**
     * 
     * @type {string}
     * @memberof TournamentSubmission
     */
    readonly _package: string;
    /**
     * 
     * @type {string}
     * @memberof TournamentSubmission
     */
    readonly description: string;
    /**
     * 
     * @type {string}
     * @memberof TournamentSubmission
     */
    readonly tournament: string;
}